requirements.txt
gradio
sentence-transformers
nltk
python-docx
numpy
